ai_image_classifier_streamlit/requirements.txt
streamlit==1.40.0
ollama==0.4.6
groq==0.33.0
pillow==11.1.0
python-dotenv==1.0.1
pandas<2.3,>=2.0